+++ sync-for-payroll/src/sdk/sdk.ts
@@ -61,9 +61,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "1.2.0";
-    genVersion = "2.159.2";
-    userAgent = "speakeasy-sdk/typescript 1.2.0 2.159.2 3.0.0 @codat/sync-for-payroll";
+    sdkVersion = "1.3.0";
+    genVersion = "2.173.0";
+    userAgent = "speakeasy-sdk/typescript 1.3.0 2.173.0 3.0.0 @codat/sync-for-payroll";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);